extract wave workflow routes
